requirements.txt

snscrape 
python-dotenv 
textblob 
transformers 
torch 
emoji 
altair 
